+++ application/comit_node/src/swap_protocols/rfc003/alice/actions/mod.rs
@@ -13,9 +13,9 @@ use crate::swap_protocols::{
 use bitcoin_support::{BitcoinQuantity, OutPoint};
 use bitcoin_witness::PrimedInput;
 use ethereum_support::{Address as EthereumAddress, Bytes, EtherQuantity};
-mod btc_erc20;
-mod btc_eth;
-mod erc20_btc;
+
+mod erc20_actions;
+mod non_erc20_actions;
 
 #[derive(Debug, Clone, PartialOrd, Ord, PartialEq, Eq, Hash)]
 pub enum ActionKind<Deploy, Fund, Redeem, Refund> {
@@ -34,24 +34,23 @@ impl<Deploy, Fund, Redeem, Refund> ActionKind<Deploy, Fund, Redeem, Refund> {
     }
 }
 
-pub trait CreateActions<AL: Ledger, AA: Asset> {
+pub trait CreateActions<L: Ledger, A: Asset> {
     type FundActionOutput;
     type RefundActionOutput;
     type RedeemActionOutput;
 
-    fn fund_action(htlc_params: HtlcParams<AL, AA>) -> Self::FundActionOutput;
+    fn fund_action(htlc_params: HtlcParams<L, A>) -> Self::FundActionOutput;
 
     fn refund_action(
-        alpha_asset: AA,
-        htlc_params: HtlcParams<AL, AA>,
-        alpha_htlc_location: AL::HtlcLocation,
+        htlc_params: HtlcParams<L, A>,
+        htlc_location: L::HtlcLocation,
         secret_source: &dyn SecretSource,
     ) -> Self::RefundActionOutput;
 
     fn redeem_action(
-        alpha_asset: AA,
-        htlc_params: HtlcParams<AL, AA>,
-        alpha_htlc_location: AL::HtlcLocation,
+        alpha_asset: A,
+        htlc_params: HtlcParams<L, A>,
+        alpha_htlc_location: L::HtlcLocation,
         secret_source: &dyn SecretSource,
         secret: Secret,
     ) -> Self::RedeemActionOutput;
@@ -73,7 +72,6 @@ impl CreateActions<Bitcoin, BitcoinQuantity> for (Bitcoin, BitcoinQuantity) {
     }
 
     fn refund_action(
-        alpha_asset: BitcoinQuantity,
         htlc_params: HtlcParams<Bitcoin, BitcoinQuantity>,
         alpha_htlc_location: OutPoint,
         secret_source: &dyn SecretSource,
@@ -83,7 +81,7 @@ impl CreateActions<Bitcoin, BitcoinQuantity> for (Bitcoin, BitcoinQuantity) {
         bitcoin::SpendOutput {
             output: PrimedInput::new(
                 alpha_htlc_location,
-                alpha_asset,
+                htlc_params.asset,
                 htlc.unlock_after_timeout(secret_source.secp256k1_refund()),
             ),
             network: htlc_params.ledger.network,
@@ -120,7 +118,6 @@ impl CreateActions<Ethereum, EtherQuantity> for (Ethereum, EtherQuantity) {
     }
 
     fn refund_action(
-        _alpha_asset: EtherQuantity,
         htlc_params: HtlcParams<Ethereum, EtherQuantity>,
         alpha_htlc_location: EthereumAddress,
         _secret_source: &dyn SecretSource,
@@ -156,166 +153,3 @@ impl CreateActions<Ethereum, EtherQuantity> for (Ethereum, EtherQuantity) {
         }
     }
 }
-
-//////// Old stuff
-// trait DeployAction<AL: Ledger, BL: Ledger, AA: Asset, BA: Asset> {
-//    type DeployActionReturn;
-//    fn deploy_action(
-//        request: &rfc003::messages::Request<AL, BL, AA, BA>,
-//        response: &rfc003::messages::AcceptResponseBody<AL, BL>,
-//    ) -> Self::DeployActionReturn;
-//}
-
-// trait FundAction<AL: Ledger, AA: Asset> {
-// type Fund;
-// fn fund_action(htlc_params: HtlcParams<AL, AA>, amount: AA, network:
-// AL::Network) -> Self::Fund;
-// }
-//
-// trait RefundAction<AL: Ledger, AA: Asset> {
-// type Output;
-//
-// fn refund_action(
-// alpha_asset: AA,
-// htlc_params: HtlcParams<AL, AA>,
-// network: AL::Network,
-// alpha_htlc_location: AL::HtlcLocation,
-// secret_source: &dyn SecretSource,
-// ) -> Self::Output;
-// }
-//
-// trait RedeemAction<AL: Ledger, BL: Ledger, AA: Asset, BA: Asset> {
-// type HTLCLocation;
-// type Output;
-//
-// fn redeem_action(
-// request: &rfc003::messages::Request<AL, BL, AA, BA>,
-// beta_htlc_location: Self::HTLCLocation,
-// secret: Secret,
-// ) -> Self::Output;
-// }
-//
-// impl<BL: Ledger, BA: Asset> DeployAction<Ethereum, BL, Erc20Token, BA> for
-// (Ethereum, Erc20Token) {    type DeployActionReturn =
-// ethereum::ContractDeploy;
-//
-//    fn deploy_action(
-//        request: &rfc003::messages::Request<Ethereum, BL, Erc20Token, BA>,
-//        response: &rfc003::messages::AcceptResponseBody<Ethereum, BL>,
-//    ) -> Self::DeployActionReturn {
-//        HtlcParams::new_alpha_params(request, response).into()
-//    }
-// }
-//
-// impl FundAction<Bitcoin, BitcoinQuantity> for (Bitcoin, BitcoinQuantity) {
-// type Fund = bitcoin::SendToAddress;
-//
-// fn fund_action(
-// htlc_params: HtlcParams<Bitcoin, BitcoinQuantity>,
-// alpha_asset: BitcoinQuantity,
-// network: BitcoinNetwork,
-// ) -> Self::Fund {
-// let to = htlc_params.compute_address();
-//
-// bitcoin::SendToAddress {
-// to,
-// amount: alpha_asset,
-// network,
-// }
-// }
-// }
-//
-// impl<BL: Ledger, BA: Asset> FundAction<Ethereum, BL, EtherQuantity, BA>
-//    for (Ethereum, EtherQuantity)
-// {
-//    type Fund = ethereum::SendTransaction;
-//
-//    fn fund_action(
-//        request: &rfc003::messages::Request<Ethereum, BL, EtherQuantity, BA>,
-//        response: &rfc003::messages::AcceptResponseBody<Ethereum, BL>,
-//    ) -> Self::Fund {
-//        let to = request.alpha_asset.token_contract;
-//        let htlc = Erc20Htlc::from(HtlcParams::new_alpha_params(request,
-// response));        let gas_limit = Erc20Htlc::fund_tx_gas_limit();
-//        let network = request.alpha_ledger.network;
-//
-//        ethereum::SendTransaction {
-//            to,
-//            data: htlc.funding_tx_payload(alpha_htlc_location),
-//            gas_limit,
-//            amount: EtherQuantity::zero(),
-//            network,
-//        }
-//    }
-// }
-//
-// impl RefundAction<Bitcoin, BitcoinQuantity> for (Bitcoin, BitcoinQuantity) {
-// type Output = bitcoin::SpendOutput;
-//
-// fn refund_action(
-// alpha_asset: BitcoinQuantity,
-// htlc_params: HtlcParams<Bitcoin, BitcoinQuantity>,
-// network: BitcoinNetwork,
-// alpha_htlc_location: OutPoint,
-// secret_source: &dyn SecretSource,
-// ) -> Self::Output {
-// let htlc = bitcoin::Htlc::from(htlc_params);
-//
-// bitcoin::SpendOutput {
-// output: PrimedInput::new(
-// alpha_htlc_location,
-// alpha_asset,
-// htlc.unlock_after_timeout(secret_source.secp256k1_refund()),
-// ),
-// network,
-// }
-// }
-// }
-//
-// impl<AL: Ledger, AA: Asset> RedeemAction<AL, Ethereum, AA, EtherQuantity>
-// for (Ethereum, EtherQuantity)
-// {
-// type HTLCLocation = ethereum_support::Address;
-// type Output = ethereum::SendTransaction;
-//
-// fn redeem_action(
-// request: &rfc003::messages::Request<AL, Ethereum, AA, EtherQuantity>,
-// beta_htlc_location: Self::HTLCLocation,
-// secret: Secret,
-// ) -> Self::Output {
-// let data = Bytes::from(secret.raw_secret().to_vec());
-// let gas_limit = EtherHtlc::tx_gas_limit();
-// let network = request.beta_ledger.network;
-//
-// ethereum::SendTransaction {
-// to: beta_htlc_location,
-// data,
-// gas_limit,
-// amount: EtherQuantity::zero(),
-// network,
-// }
-// }
-// }
-//
-// impl<AL: Ledger, AA: Asset> RedeemAction<AL, Ethereum, AA, Erc20Token> for
-// (Ethereum, Erc20Token) { type HTLCLocation = ethereum_support::Address;
-// type Output = ethereum::SendTransaction;
-//
-// fn redeem_action(
-// request: &rfc003::messages::Request<AL, Ethereum, AA, Erc20Token>,
-// beta_htlc_location: Self::HTLCLocation,
-// secret: Secret,
-// ) -> Self::Output {
-// let data = Bytes::from(secret.raw_secret().to_vec());
-// let gas_limit = Erc20Htlc::tx_gas_limit();
-// let network = request.beta_ledger.network;
-//
-// ethereum::SendTransaction {
-// to: beta_htlc_location,
-// data,
-// gas_limit,
-// amount: EtherQuantity::zero(),
-// network,
-// }
-// }
-// }
